agents-backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
langchain
langchain-openai
python-dotenv
colorama
